Rename matlab_solver.py to matlab_interface.py for better architectural clarity
- Rename matlab_solver.py to matlab_interface.py to better reflect its role as an interface between Python and MATLAB systems
- Update all import statements across codebase to use new module name
- Update documentation references and examples
- Update test files and mock patches to use new module name
- Update logger name in module for consistency
- This naming better distinguishes the file's role as a system integration component rather than a direct solver implementation
