+++ tests/test_matlab_solver_enhanced.py
@@ -15,7 +15,7 @@ from pathlib import Path
 project_root = Path(__file__).parent.parent
 sys.path.insert(0, str(project_root))
 
-from scripts.solvers.matlab_octave.matlab_solver import MatlabSolver
+from scripts.solvers.matlab_octave.matlab_interface import MatlabSolver
 from scripts.data_loaders.problem_loader import ProblemData
 from scripts.utils.logger import get_logger
